refactor loading of 'section runners'
instead of calling *-run-section on each runner section, we now create a
defrecord'ed service object, which we call start on.

this decouples ini file parsing from starting the service and may allow
us to define more methods on the service (stop comes to mind).

it also gives services the opportunity to store some service related
information in a defrecord.